Removed update_config and added copy_to_file methods
pyimagediet now has function with same functionality as update_config so
we should use that one.
copy_to_file has been added because uploaded files may not even be
stored on a filesystem, but writeable directory is needed for external
tools to work. This method copies file to that (configured) directory.